data_init/pyproject.toml
[tool.poetry]
name = "data_init"
version = "0.1.0"
description = ""
authors = ["FatMadLad <mrammler@web.de>"]
packages = [{include = "data_init"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
boto3 = "^1.26.115"
tqdm = "^4.65.0"
numpy = "^1.24.3"
pillow = "^9.5.0"
sentence-transformers = "^2.2.2"
qdrant-client = "^1.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
